LPS-97669 | LPS-103897 Move test base to util package
