+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-test/src/testIntegration/resources/com/liferay/portal/workflow/metrics/service/util/dependencies/single-approver-definition.xml
